Move public Protocol definitions from .pyi to .py
