+++ docs/docs/api/storages.md
@@ -1,5 +1,7 @@
 # rustuna.storages
 
+::: rustuna.storages.StorageProtocol
+
 ::: rustuna.storages.InMemoryStorage
 
 ::: rustuna.storages.SQLite3Storage
