+++ docs/docs/api/trial_queue.md
@@ -1,5 +1,7 @@
 # rustuna.trial_queue
 
+::: rustuna.trial_queue.TrialQueueProtocol
+
 ::: rustuna.trial_queue.InMemoryTrialQueue
 
 ::: rustuna.trial_queue.SQLite3TrialQueue
