+++ rustuna_pyo3/rustuna/samplers.py
@@ -1,10 +1,6 @@
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from rustuna._rustuna import SamplerProtocol
-
+from rustuna._protocols import SamplerProtocol
 from rustuna._rustuna import Sampler, SamplerContext
 
 __all__ = [
